@@ -9,107 +9,75 @@ namespace USPS;
  * @since  1.0
  *
  * @author Vincent Gabriel
+ * @author Shivam Mathur
  */
 class Address
 {
     /**
      * @var array list of all key=>value pairs we added so far for the current address
      */
-    protected $addressInfo = [];
+    protected array $addressInfo = [];
 
     /**
      * Set the address2 property.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setAddress($value)
+    public function setAddress(string|int $value): self
     {
         return $this->setField('Address2', $value);
     }
 
     /**
      * Set the address1 property usually the apt or suit number.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setApt($value)
+    public function setApt(string|int $value): self
     {
         return $this->setField('Address1', $value);
     }
 
     /**
      * Set the city property.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setCity($value)
+    public function setCity(string|int $value): self
     {
         return $this->setField('City', $value);
     }
 
     /**
      * Set the state property.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setState($value)
+    public function setState(string|int $value): self
     {
         return $this->setField('State', $value);
     }
 
     /**
      * Set the zip4 property - zip code value represented by 4 integers.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setZip4($value)
+    public function setZip4(string|int $value): self
     {
         return $this->setField('Zip4', $value);
     }
 
     /**
      * Set the zip5 property - zip code value represented by 5 integers.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setZip5($value)
+    public function setZip5(string|int $value): self
     {
         return $this->setField('Zip5', $value);
     }
 
     /**
      * Set the firmname property.
-     *
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setFirmName($value)
+    public function setFirmName(string|int $value): self
     {
         return $this->setField('FirmName', $value);
     }
 
     /**
      * Add an element to the stack.
-     *
-     * @param string|int $key
-     * @param string|int $value
-     *
-     * @return object Address
      */
-    public function setField($key, $value)
+    public function setField(string|int $key, mixed $value): self
     {
         $this->addressInfo[ucwords($key)] = $value;
 
@@ -118,10 +86,8 @@ class Address
 
     /**
      * Returns a list of all the info we gathered so far in the current address object.
-     *
-     * @return array
      */
-    public function getAddressInfo()
+    public function getAddressInfo(): array
     {
         return $this->addressInfo;
     }
