@@ -2,26 +2,27 @@
 
 namespace USPS;
 
+use Exception;
+
 /**
  * Class OpenDistributeLabel.
  */
 class OpenDistributeLabel extends USPSBase
 {
     /**
-     * @var string - the api version used for this type of call
+     * the api version used for this type of call
      */
-    protected $apiVersion = 'OpenDistributePriorityV2';
+    protected string $apiVersion = 'OpenDistributePriorityV2';
     /**
-     * @var array - route added so far.
+     * route added so far.
      */
-    protected $fields = [];
+    protected array $fields = [];
 
     /**
      * Perform the API call.
-     *
-     * @return string
+     * @throws Exception
      */
-    public function createLabel()
+    public function createLabel(): bool|string
     {
         // Add missing required
         $this->addMissingRequired();
@@ -45,17 +46,13 @@ class OpenDistributeLabel extends USPSBase
 
     /**
      * Return the USPS confirmation/tracking number if we have one.
-     *
-     * @return string|bool
      */
-    public function getConfirmationNumber()
+    public function getConfirmationNumber(): bool|string
     {
         $response = $this->getArrayResponse();
         // Check to make sure we have it
-        if (isset($response[$this->getResponseApiName()])) {
-            if (isset($response[$this->getResponseApiName()]['OpenDistributePriorityNumber'])) {
-                return $response[$this->getResponseApiName()]['OpenDistributePriorityNumber'];
-            }
+        if (isset($response[$this->getResponseApiName()]) && isset($response[$this->getResponseApiName()]['OpenDistributePriorityNumber'])) {
+            return $response[$this->getResponseApiName()]['OpenDistributePriorityNumber'];
         }
 
         return false;
@@ -63,18 +60,14 @@ class OpenDistributeLabel extends USPSBase
 
     /**
      * Return the USPS label as a base64 encoded string.
-     *
-     * @return string|bool
      */
-    public function getLabelContents()
+    public function getLabelContents(): bool|string
     {
         $response = $this->getArrayResponse();
 
         // Check to make sure we have it
-        if (isset($response[$this->getResponseApiName()])) {
-            if (isset($response[$this->getResponseApiName()]['OpenDistributePriorityLabel'])) {
-                return $response[$this->getResponseApiName()]['OpenDistributePriorityLabel'];
-            }
+        if (isset($response[$this->getResponseApiName()]) && isset($response[$this->getResponseApiName()]['OpenDistributePriorityLabel'])) {
+            return $response[$this->getResponseApiName()]['OpenDistributePriorityLabel'];
         }
 
         return false;
@@ -83,39 +76,26 @@ class OpenDistributeLabel extends USPSBase
     /**
      * returns array of all fields added.
      *
-     * @return array
      */
-    public function getPostFields()
+    public function getPostFields(): array
     {
         return $this->fields;
     }
 
     /**
      * Set the from address.
-     *
-     * @param string $firstName
-     * @param string $lastName
-     * @param string $company
-     * @param string $address
-     * @param string $city
-     * @param string $state
-     * @param string $zip
-     * @param string $address2
-     * @param string $zip4
-     *
-     * @return object
      */
     public function setFromAddress(
-        $firstName,
-        $lastName,
-        $company,
-        $address,
-        $city,
-        $state,
-        $zip,
-        $address2 = null,
-        $zip4 = null
-    ) {
+        string $firstName,
+        string $lastName,
+        string $company,
+        string $address,
+        string $city,
+        string $state,
+        string $zip,
+        string|null $address2 = null,
+        string|null $zip4 = null
+    ): self {
         $this->setField(5, 'FromName', trim($firstName.' '.$lastName));
         $this->setField(7, 'FromFirm', $company);
         $this->setField(8, 'FromAddress1', $address2);
@@ -130,18 +110,16 @@ class OpenDistributeLabel extends USPSBase
 
     /**
      * Set the to address.
-     *
-     * @param string $company
-     * @param string $address
-     * @param string $city
-     * @param string $state
-     * @param string $zip
-     * @param string $address2
-     * @param string $zip4
-     *
-     * @return object
      */
-    public function setToAddress($company, $address, $city, $state, $zip, $address2 = null, $zip4 = null)
+    public function setToAddress(
+        string $company,
+        string $address,
+        string $city,
+        string $state,
+        string $zip,
+        string|null $address2 = null,
+        string|null $zip4 = null
+    ): self
     {
         $this->setField(15, 'ToFacilityName', $company);
         $this->setField(18, 'ToFacilityAddress1', $address2);
@@ -156,12 +134,8 @@ class OpenDistributeLabel extends USPSBase
 
     /**
      * Set package weight in ounces.
-     *
-     * @param $weight
-     *
-     * @return $this
      */
-    public function setWeightOunces($weight)
+    public function setWeightOunces(string $weight): self
     {
         $this->setField(38, 'WeightInOunces', $weight);
 
@@ -170,12 +144,8 @@ class OpenDistributeLabel extends USPSBase
 
     /**
      * Set package weight in ounces.
-     *
-     * @param $weight
-     *
-     * @return $this
      */
-    public function setWeightPounds($weight)
+    public function setWeightPounds(mixed $weight): self
     {
         $this->setField(37, 'WeightInPounds', $weight);
 
@@ -183,16 +153,10 @@ class OpenDistributeLabel extends USPSBase
     }
 
     /**
-     * Set any other requried string make sure you set the correct position as well
+     * Set any other required string make sure you set the correct position as well
      * as the position of the items matters.
-     *
-     * @param int    $position
-     * @param string $key
-     * @param string $value
-     *
-     * @return object
      */
-    public function setField($position, $key, $value)
+    public function setField(int $position, string $key, string $value): self
     {
         $this->fields[$position.':'.$key] = $value;
 
@@ -201,10 +165,8 @@ class OpenDistributeLabel extends USPSBase
 
     /**
      * Add missing required elements.
-     *
-     * @return void
      */
-    protected function addMissingRequired()
+    protected function addMissingRequired(): void
     {
         $required = [
             '1:Revision'                      => '',
