@@ -2,26 +2,27 @@
 
 namespace USPS;
 
+use Exception;
+
 /**
  * Class PriorityLabel.
  */
 class PriorityLabel extends USPSBase
 {
     /**
-     * @var string - the api version used for this type of call
+     * the api version used for this type of call
      */
-    protected $apiVersion = 'ExpressMailLabel';
+    protected string $apiVersion = 'ExpressMailLabel';
     /**
-     * @var array - route added so far.
+     * route added so far.
      */
-    protected $fields = [];
+    protected array $fields = [];
 
     /**
      * Perform the API call.
-     *
-     * @return string
+     * @throws Exception
      */
-    public function createLabel()
+    public function createLabel(): bool|string
     {
         // Add missing required
         $this->addMissingRequired();
@@ -45,17 +46,13 @@ class PriorityLabel extends USPSBase
 
     /**
      * Return the USPS confirmation/tracking number if we have one.
-     *
-     * @return string|bool
      */
-    public function getConfirmationNumber()
+    public function getConfirmationNumber(): bool|string
     {
         $response = $this->getArrayResponse();
         // Check to make sure we have it
-        if (isset($response[$this->getResponseApiName()])) {
-            if (isset($response[$this->getResponseApiName()]['EMConfirmationNumber'])) {
-                return $response[$this->getResponseApiName()]['EMConfirmationNumber'];
-            }
+        if (isset($response[$this->getResponseApiName()]) && isset($response[$this->getResponseApiName()]['EMConfirmationNumber'])) {
+            return $response[$this->getResponseApiName()]['EMConfirmationNumber'];
         }
 
         return false;
@@ -63,17 +60,13 @@ class PriorityLabel extends USPSBase
 
     /**
      * Return the USPS label as a base64 encoded string.
-     *
-     * @return string|bool
      */
-    public function getLabelContents()
+    public function getLabelContents(): bool|string
     {
         $response = $this->getArrayResponse();
         // Check to make sure we have it
-        if (isset($response[$this->getResponseApiName()])) {
-            if (isset($response[$this->getResponseApiName()]['EMLabel'])) {
-                return $response[$this->getResponseApiName()]['EMLabel'];
-            }
+        if (isset($response[$this->getResponseApiName()]) && isset($response[$this->getResponseApiName()]['EMLabel'])) {
+            return $response[$this->getResponseApiName()]['EMLabel'];
         }
 
         return false;
@@ -81,17 +74,13 @@ class PriorityLabel extends USPSBase
 
     /**
      * Return the USPS receipt as a base64 encoded string.
-     *
-     * @return string|bool
      */
-    public function getReceiptContents()
+    public function getReceiptContents(): bool|string
     {
         $response = $this->getArrayResponse();
         // Check to make sure we have it
-        if (isset($response[$this->getResponseApiName()])) {
-            if (isset($response[$this->getResponseApiName()]['EMReceipt'])) {
-                return $response[$this->getResponseApiName()]['EMReceipt'];
-            }
+        if (isset($response[$this->getResponseApiName()]) && isset($response[$this->getResponseApiName()]['EMReceipt'])) {
+            return $response[$this->getResponseApiName()]['EMReceipt'];
         }
 
         return false;
@@ -100,41 +89,27 @@ class PriorityLabel extends USPSBase
     /**
      * returns array of all fields added.
      *
-     * @return array
      */
-    public function getPostFields()
+    public function getPostFields(): array
     {
         return $this->fields;
     }
 
     /**
      * Set the from address.
-     *
-     * @param string $firstName
-     * @param string $lastName
-     * @param string $company
-     * @param string $address
-     * @param string $city
-     * @param string $state
-     * @param string $zip
-     * @param string $address2
-     * @param string $zip4
-     * @param string $phone
-     *
-     * @return object
      */
     public function setFromAddress(
-        $firstName,
-        $lastName,
-        $company,
-        $address,
-        $city,
-        $state,
-        $zip,
-        $address2 = null,
-        $zip4 = null,
-        $phone = null
-    ) {
+        string $firstName,
+        string $lastName,
+        string $company,
+        string $address,
+        string $city,
+        string $state,
+        string $zip,
+        string|null $address2 = null,
+        string|null $zip4 = null,
+        string|null $phone = null
+    ): self {
         $this->setField(5, 'FromFirstName', $firstName);
         $this->setField(6, 'FromLastName', $lastName);
         $this->setField(7, 'FromFirm', $company);
@@ -151,32 +126,19 @@ class PriorityLabel extends USPSBase
 
     /**
      * Set the to address.
-     *
-     * @param string $firstName
-     * @param string $lastName
-     * @param string $company
-     * @param string $address
-     * @param string $city
-     * @param string $state
-     * @param string $zip
-     * @param string $address2
-     * @param string $zip4
-     * @param string $phone
-     *
-     * @return object
      */
     public function setToAddress(
-        $firstName,
-        $lastName,
-        $company,
-        $address,
-        $city,
-        $state,
-        $zip,
-        $address2 = null,
-        $zip4 = null,
-        $phone = null
-    ) {
+        string $firstName,
+        string $lastName,
+        string $company,
+        string $address,
+        string $city,
+        string $state,
+        string $zip,
+        string|null $address2 = null,
+        string|null $zip4 = null,
+        string|null $phone = null
+    ): self {
         $this->setField(15, 'ToFirstName', $firstName);
         $this->setField(16, 'ToLastName', $lastName);
         $this->setField(17, 'ToFirm', $company);
@@ -193,12 +155,8 @@ class PriorityLabel extends USPSBase
 
     /**
      * Set package weight in ounces.
-     *
-     * @param $weight
-     *
-     * @return $this
      */
-    public function setWeightOunces($weight)
+    public function setWeightOunces(mixed $weight): self
     {
         $this->setField(25, 'WeightInOunces', $weight);
 
@@ -206,16 +164,10 @@ class PriorityLabel extends USPSBase
     }
 
     /**
-     * Set any other requried string make sure you set the correct position as well
+     * Set any other required string make sure you set the correct position as well
      * as the position of the items matters.
-     *
-     * @param int    $position
-     * @param string $key
-     * @param string $value
-     *
-     * @return object
      */
-    public function setField($position, $key, $value)
+    public function setField(int $position, string $key, string $value): self
     {
         $this->fields[$position.':'.$key] = $value;
 
@@ -224,10 +176,8 @@ class PriorityLabel extends USPSBase
 
     /**
      * Add missing required elements.
-     *
-     * @return void
      */
-    protected function addMissingRequired()
+    protected function addMissingRequired(): void
     {
         $required = [
             '1:Option'                 => '',
