@@ -1,17 +1,20 @@
 <?php
 
+use USPS\Address;
+use USPS\AddressVerify;
+
 require_once('autoload.php');
 
 // Initiate and set the username provided from usps
-$verify = new \USPS\AddressVerify('xxxx');
+$verify = new AddressVerify('xxxx');
 
 // During test mode this seems not to always work as expected
 //$verify->setTestMode(true);
 
 // Create new address object and assign the properties
-// apartently the order you assign them is important so make sure
+// apparently the order you assign them is important so make sure
 // to set them as the example below
-$address = new \USPS\Address();
+$address = new Address();
 $address->setFirmName('Apartment');
 $address->setApt('100');
 $address->setAddress('9200 Milliken Ave');
@@ -24,7 +27,11 @@ $address->setZip4('');
 $verify->addAddress($address);
 
 // Perform the request and return result
-print_r($verify->verify());
+try {
+    print_r($verify->verify());
+} catch (Exception $e) {
+    // Handle any errors
+}
 print_r($verify->getArrayResponse());
 
 var_dump($verify->isError());
