@@ -1,17 +1,20 @@
 <?php
 
+use USPS\Address;
+use USPS\ZipCodeLookup;
+
 require_once('autoload.php');
 
 // Initiate and set the username provided from usps
-$zipcode = new \USPS\ZipCodeLookup('xxxx');
+$zipcode = new ZipCodeLookup('xxxx');
 
 // During test mode this seems not to always work as expected
 //$verify->setTestMode(true);
 
 // Create new address object and assign the properties
-// apartently the order you assign them is important so make sure
+// apparently the order you assign them is important so make sure
 // to set them as the example below
-$address = new \USPS\Address();
+$address = new Address();
 $address->setFirmName('Apartment');
 $address->setApt('100');
 $address->setAddress('9200 Milliken Ave');
@@ -22,7 +25,11 @@ $address->setState('CA');
 $zipcode->addAddress($address);
 
 // Perform the call and print out the results
-print_r($zipcode->lookup());
+try {
+    print_r($zipcode->lookup());
+} catch (Exception $e) {
+    // Handle any errors
+}
 print_r($zipcode->getArrayResponse());
 
 // Check if it was completed
